# pes-spring/src/main/resources/openapi/pes.yaml
openapi: 3.0.3
info:
  title: Kaspian PES application API
  description: "kaspian: PES application API Documentation"
  termsOfService: http://kaspian.sk/terms/
  contact:
    email: kaspian@kaspian.sk
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "1.0"
...
servers:
  - url: http://kapian-pes.online/
...
paths:
  /api/card/save:
    post: 
      security:
        - bearerAuth: []
      tags: 
        - Cards
      summary: Save object type of Card
      operationId: saveCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Card'
      responses:
        201:
          description: "Successfully saved Card"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/card/{code}:
    get: 
      security:
        - bearerAuth: []
      tags: 
        - Cards
      summary: Return object type of Card base on code
      operationId: getCard
      parameters:
        - in: path
          name: code
          description: String code of card
          required: true
          schema:
            type: string
      responses:
        200:
          description: "Successfully get Card"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
    delete:
      security:
        - bearerAuth: []
      tags: 
        - Cards
      summary: Remove object type of Card
      operationId: deleteCard
      parameters:
        - in: path
          name: code
          description: String code of card
          required: true
          schema:
            type: string
      responses:
        204:
          description: "Successfully removed Card"
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/card/allCards:
    get:
      security:
        - bearerAuth: []
      tags: 
        - Cards
      summary: Return all cards
      operationId: getAllCard
      responses:
        200:
          description: "Successfully get All Cards"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/card/activate/{code}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Cards
      summary: Endpoint change/ toggle status of activity on object
      operationId: getToggleCard
      parameters:
        - in: path
          name: code
          description: String code of card
          required: true
          schema:
            type: string
      responses:
        200:
          description: "Successfully get Card"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/entry/save:
    post:
      security:
        - bearerAuth: []
      tags: 
        - Entrys
      summary: Save object type of Entry
      operationId: saveEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Entry'
      responses:
        201:
          description: "Successfully saved Entry"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/entry/{code}:
    get:
      security:
        - bearerAuth: []
      tags: 
        - Entrys
      summary: Return object type of Entry base on code
      operationId: getEntry
      parameters:
        - in: path
          name: code
          description: String code of entry
          required: true
          schema:
            type: string
      responses:
        200:
          description: "Successfully get Entry"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
    delete:
      security:
        - bearerAuth: []
      tags: 
        - Entrys
      summary: Remove object type of Entry
      operationId: deleteEntry
      parameters:
        - in: path
          name: code
          description: String code of entry
          required: true
          schema:
            type: string
      responses:
        204:
          description: "Successfully removed Entry"
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/entry/allEntrys:
    get:
      security:
        - bearerAuth: []
      tags: 
        - Entrys
      summary: Return all entrys
      operationId: getAllEntry
      responses:
        200:
          description: "Successfully get All Entrys"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entry'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/entry/allCardEntrysMain:
    post:
      security:
        - bearerAuth: []
      tags: 
        - Events
      summary: Return all entrys for selected card
      operationId: getAllCardEntrysMain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Card'
      responses:
        200:
          description: "Successfully get All cards entrys"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entry'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/entry/allCardEntrysOnSite:
    post:
      security:
        - bearerAuth: []
      tags:
        - Events
      summary: Return all entrys for selected card
      operationId: getallCardEntrysOnSite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Card'
      responses:
        200:
          description: "Successfully get All cards entrys"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entry'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/entry/newUserEntrys:
    get:
      security:
        - bearerAuth: []
      tags: 
        - Events
      summary: Return 5 newest events for user
      operationId: getNewUserEntrys
      responses:
        200:
          description: "Successfully get All entrys"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entry'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/entry/getListNewEntrys:
            get:
              security:
                - bearerAuth: [ ]
              tags:
                - Events
              summary: Return 5 newest events for user
              operationId: getListNewEntrys
              parameters:
                - $ref: '#/components/parameters/PageSize'
              responses:
                200:
                  description: "Successfully get All entrys"
                  content:
                    application/json:
                      schema:
                        type: array
                        items:
                          $ref: '#/components/schemas/Entry'
                400:
                  $ref: '#/components/responses/InvalidRequest'
                401:
                  $ref: '#/components/responses/UnauthorizedError'
                403:
                  $ref: '#/components/responses/AccessDenied'
                404:
                  $ref: '#/components/responses/URLNotExist'
                500:
                  $ref: '#/components/responses/SystemException'
  /api/event/save:
    post:
      security:
        - bearerAuth: []
      tags: 
        - Events
      summary: Save object type of Event
      operationId: saveEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        201:
          description: "Successfully saved Event"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/event/{code}:
    get: 
      security:
        - bearerAuth: []
      tags: 
        - Events
      summary: Return object type of Event base on code
      operationId: getEvent
      parameters:
        - in: path
          name: code
          description: String code of event
          required: true
          schema:
            type: string
      responses:
        200:
          description: "Successfully get Event"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
    delete:
      security:
        - bearerAuth: []
      tags: 
        - Events
      summary: Remove object type of Event
      operationId: deleteEvent
      parameters:
        - in: path
          name: code
          description: String code of event
          required: true
          schema:
            type: string
      responses:
        204:
          description: "Successfully removed Event"
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/event/allEvents:
    get:
      security:
        - bearerAuth: []
      tags: 
        - Events
      summary: Return all events
      operationId: getAllEvent
      responses:
        200:
          description: "Successfully get All Events"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/campaign/save:
    post:
      security:
        - bearerAuth: []
      tags: 
        - Campaigns
      summary: Save object type of Campaign
      operationId: saveCampaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Campaign'
      responses:
        201:
          description: "Successfully saved Campaign"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/campaign/{code}:
    get:
      security:
        - bearerAuth: []
      tags: 
        - Campaigns
      summary: Return object type of Campaign base on code
      operationId: getCampaign
      parameters:
        - in: path
          name: code
          description: String code of campaign
          required: true
          schema:
            type: string
      responses:
        200:
          description: "Successfully get Campaign"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
    delete:
      security:
        - bearerAuth: []
      tags: 
        - Campaigns
      summary: Remove object type of Campaign
      operationId: deleteCampaign
      parameters:
        - in: path
          name: code
          description: String code of campaign
          required: true
          schema:
            type: string
      responses:
        204:
          description: "Successfully removed Campaign"
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/campaign/allCampaigns:
    get:
      security:
        - bearerAuth: []
      tags: 
        - Campaigns
      summary: Return all campaigns
      operationId: getAllCampaign
      responses:
        200:
          description: "Successfully get All Campaigns"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Campaign'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/coworker/save:
    post:
      security:
        - bearerAuth: []
      tags: 
        - Coworkers
      summary: Save object type of Coworker
      operationId: saveCoworker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Coworker'
      responses:
        201:
          description: "Successfully saved Coworker"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coworker'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/coworker/{code}:
    get:
      security:
        - bearerAuth: []
      tags: 
        - Coworkers
      summary: Return object type of Coworker base on code
      operationId: getCoworker
      parameters:
        - in: path
          name: code
          description: String code of coworker
          required: true
          schema:
            type: string
      responses:
        200:
          description: "Successfully get Coworker"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coworker'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
    delete:
      security:
        - bearerAuth: []
      tags: 
        - Coworkers
      summary: Remove object type of Coworker
      operationId: deleteCoworker
      parameters:
        - in: path
          name: code
          description: String code of coworker
          required: true
          schema:
            type: string
      responses:
        204:
          description: "Successfully removed Coworker"
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
  /api/coworker/allCoworkers:
    get:
      security:
        - bearerAuth: []
      tags: 
        - Coworkers
      summary: Return all coworkers
      operationId: getAllCoworker
      responses:
        200:
          description: "Successfully get All Coworkers"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Coworker'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
...
  /api/institution/allInstitutions:
    get:
      security:
        - bearerAuth: []
      tags: 
        - Institutions
      summary: Return all institutions
      operationId: getAllInstitutions
      responses:
        200:
          description: "Successfully get All Institutions"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Institution'
        400:
          $ref: '#/components/responses/InvalidRequest'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/AccessDenied'
        404:
          $ref: '#/components/responses/URLNotExist'
        500:
          $ref: '#/components/responses/SystemException'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    authSessionToken:
      type: apiKey
      in: header
      name: X-Auth-Token
      description: Include the `X-Auth-Token` header for session identification.
  parameters:
    PageSize:
      name: page_size
      in: query
      description: Number of items in a page, default page size is `20`
      x-pageable: true
      x-default-value: 20
      schema:
        type: integer
        #        minimum: 1
        #        maximum: 100
        default: 20
      required: false
  responses:
...
    SystemException:
      description: "A system error occurred in the service call"
    InvalidRequest:
      description: "The request does not match the specification"
    URLNotExist:
      description: "Given URL does not exist"
    UnauthorizedError:
      description: "Access token is missing or invalid"
    AccessDenied:
      description: "Access denied - insufficient rights"
...